Enables direct messaging for slack notifications.  Also tidies up the folder structure and moves some logging to lower priorities.
