--- src/main/resources/jenkins/plugins/slack/SlackNotifier/SlackJobProperty/index.jelly
@@ -1,1 +1,0 @@
-<div></div>
