+++ src/main/resources/jenkins/plugins/slack/SlackUserProperty/config.jelly
@@ -1,0 +1,7 @@
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+
+    <f:entry title="Slack Username" help="${rootURL}/plugin/slack/help-userConfig-slackUsername.html">
+        <f:textbox name="slackUsername" value="${instance.getUsername()}" />
+    </f:entry>
+
+</j:jelly>
