+++ src/main/webapp/help-projectConfig-slackDirectMessage.html
@@ -1,0 +1,6 @@
+<div>
+    <p>
+        Check this box to have the build notifications sent directly to you. You can choose to have these notifications sent to the channel only (default behavior), to both you and the channel, or to just you.<br>
+        <b><u>You must add your Slack username from the User Configuration page to enable direct messaging!</u></b>
+    </p>
+</div>
