+++ src/main/webapp/help-userConfig-slackUsername.html
@@ -1,0 +1,5 @@
+<div>
+    <p>
+        This is the username that you use on Slack. This must be set in order to receive direct messages.
+    </p>
+</div>
